Add test that reproduces bug
