+++ src/composables/queries/__mocks__/useGaugeVotesQuery.ts
@@ -1,0 +1,7 @@
+export default function useGaugeVotesQuery() {
+  return {
+    data: [],
+    isLoading: false,
+    refetch: jest.fn()
+  };
+}
